breaking: deprecate fw_utils.parse_s3_url in favor of s3.parse_s3_url
