+++ tests/unit/test_s3.py
@@ -16,7 +16,7 @@ import os
 import pytest
 from mock import Mock
 
-from sagemaker.s3 import S3Uploader, S3Downloader
+from sagemaker import s3
 
 BUCKET_NAME = "mybucket"
 REGION = "us-west-2"
@@ -42,7 +42,7 @@ def sagemaker_session():
 
 def test_upload(sagemaker_session, caplog):
     desired_s3_uri = os.path.join("s3://", BUCKET_NAME, CURRENT_JOB_NAME, SOURCE_NAME)
-    S3Uploader.upload(
+    s3.S3Uploader.upload(
         local_path="/path/to/app.jar",
         desired_s3_uri=desired_s3_uri,
         sagemaker_session=sagemaker_session,
@@ -57,7 +57,7 @@ def test_upload(sagemaker_session, caplog):
 
 def test_upload_with_kms_key(sagemaker_session):
     desired_s3_uri = os.path.join("s3://", BUCKET_NAME, CURRENT_JOB_NAME, SOURCE_NAME)
-    S3Uploader.upload(
+    s3.S3Uploader.upload(
         local_path="/path/to/app.jar",
         desired_s3_uri=desired_s3_uri,
         kms_key=KMS_KEY,
@@ -73,7 +73,7 @@ def test_upload_with_kms_key(sagemaker_session):
 
 def test_download(sagemaker_session):
     s3_uri = os.path.join("s3://", BUCKET_NAME, CURRENT_JOB_NAME, SOURCE_NAME)
-    S3Downloader.download(
+    s3.S3Downloader.download(
         s3_uri=s3_uri, local_path="/path/for/download/", sagemaker_session=sagemaker_session
     )
     sagemaker_session.download_data.assert_called_with(
@@ -86,7 +86,7 @@ def test_download(sagemaker_session):
 
 def test_download_with_kms_key(sagemaker_session):
     s3_uri = os.path.join("s3://", BUCKET_NAME, CURRENT_JOB_NAME, SOURCE_NAME)
-    S3Downloader.download(
+    s3.S3Downloader.download(
         s3_uri=s3_uri,
         local_path="/path/for/download/",
         kms_key=KMS_KEY,
@@ -98,3 +98,15 @@ def test_download_with_kms_key(sagemaker_session):
         key_prefix=os.path.join(CURRENT_JOB_NAME, SOURCE_NAME),
         extra_args={"SSECustomerKey": KMS_KEY},
     )
+
+
+def test_parse_s3_url():
+    bucket, key_prefix = s3.parse_s3_url("s3://bucket/code_location")
+    assert "bucket" == bucket
+    assert "code_location" == key_prefix
+
+
+def test_parse_s3_url_fail():
+    with pytest.raises(ValueError) as error:
+        s3.parse_s3_url("t3://code_location")
+    assert "Expecting 's3' scheme" in str(error)
